Add files via upload for the metal chassis robot
This is our update code up to this point where the metal chassis robot is in the starting box, driving forward a litte, turning right and going half-way down the middle black line to grab poms.  It will put its arm down to block in poms and back up to prepare to place poms in Isolation Area.